+++ tiled/client/__init__.py
@@ -3,3 +3,4 @@ from .constructors import from_context, from_profile, from_uri  # noqa: F401
 from .container import ASCENDING, DESCENDING  # noqa: F401
 from .context import Context  # noqa: F401
 from .logger import hide_logs, record_history, show_logs  # noqa: F401
+from .metadata_update import DELETE_KEY  # noqa: F401